complete manager deletion functionality, change deletion in CRUD

+++ CRUD.py
@@ -71,13 +71,16 @@ def show_teams(name):
     #Delete Operators for player and managers
 
 def delete_player(players_id):
-    query = session.query(Player).where(id == players_id)
+    query = session.query(Player).where(id == players_id).first()
 
     session.delete(query)
     session.commit()
 
 def delete_manager(managers_id):
-    query = session.query(Manager).where(id==managers_id)
+    query = session.query(Manager).where(Manager.id==managers_id).one()
+
+    if query is None:
+        raise Exception (f"{managers_id } does not exist")
 
     session.delete(query)
     session.commit()
